加入了权限控制

+++ BusReg/src/main/java/com/ww/busReg/controller/UimController.java
@@ -10,6 +10,7 @@ import java.util.Map;
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
@@ -17,8 +18,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.ww.busReg.domain.BusinessHall;
 import com.ww.busReg.domain.Suboffice;
 import com.ww.busReg.domain.Uim;
+import com.ww.busReg.domain.User;
+import com.ww.busReg.service.BusinessHallService;
 import com.ww.busReg.service.SubofficeService;
 import com.ww.busReg.service.UimService;
 import com.ww.busReg.util.SysConfig;
@@ -33,13 +37,16 @@ public class UimController {
 	UimService uimService;
 	@Resource
 	SubofficeService subofficeService;
+	@Resource
+	BusinessHallService businessHallService;
 	@RequestMapping("/add")
-	public String addUim(Uim uim,@RequestParam MultipartFile cardPhotoFrontF,@RequestParam MultipartFile cardPhotoBackF,@RequestParam MultipartFile userCardVoideoF,@RequestParam MultipartFile userCardPhotoF,HttpServletRequest request) throws IllegalStateException, IOException
+	public String addUim(Uim uim,@RequestParam MultipartFile cardPhotoFrontF,@RequestParam MultipartFile cardPhotoBackF,@RequestParam MultipartFile userCardVoideoF,@RequestParam MultipartFile userCardPhotoF,HttpServletRequest request,HttpSession session) throws IllegalStateException, IOException
 	{	
+		User user = (User) session.getAttribute("user");
 		String filePre= SysUtil.getFilePre(uim.getCardNum());
 		
-		uim.setSubofficeId(0);
-		uim.setBusinessHallId(0);
+		uim.setSubofficeId(user.getSubofficeId());
+		uim.setBusinessHallId(user.getBusinessHallId());
 		uim.setDealState(0);
 		
 		String fileName=filePre+cardPhotoFrontF.getName()+SysUtil.getSuffix(cardPhotoFrontF.getOriginalFilename());
@@ -78,41 +85,94 @@ public class UimController {
 //		return "workBill/uim/list1";
 //	}
 	@RequestMapping("/list")
-	public String listDetails(String userName,String cardNum,Integer subofficeId,Integer businessHallId, Integer currentPage,HttpServletRequest request)
+	public String listDetails(String userName,String cardNum,Integer subofficeId,Integer businessHallId, Integer currentPage,HttpServletRequest request,HttpSession session)
+	{
+		request.setAttribute("subofficeId", subofficeId);
+		request.setAttribute("businessHallId", businessHallId);	
+		request.setAttribute("userName", userName);
+		request.setAttribute("cardNum", cardNum);
+		User user = (User) session.getAttribute("user");
+		switch (user.getLimitsId()) {
+		case 1:
+			return list1(userName,cardNum,subofficeId,businessHallId,currentPage,request);				
+		case 2:
+			return list2(userName,cardNum,businessHallId,currentPage,request,user);			
+		case 3:
+			return list3(userName,cardNum,currentPage,request,user);
+		default:
+			break;
+		}
+		return "";
+	}
+	private String list3(String userName, String cardNum, Integer currentPage,
+			HttpServletRequest request, User user) {
+		PageResout pageResout=new PageResout();
+		pageResout.setCurrentPage(currentPage);
+		uimService.findForPageForShow(userName,cardNum,user.getSubofficeId(),user.getBusinessHallId(),pageResout);
+		request.setAttribute("pageResout", pageResout);
+		return "workBill/uim/list3";
+	}
+	private String 	list1(String userName,String cardNum,Integer subofficeId,Integer businessHallId, Integer currentPage,HttpServletRequest request)
 	{
 		List<Suboffice> suboffices= subofficeService.getAllUsable();
 		request.setAttribute("suboffices", suboffices);
-		request.setAttribute("subofficeId", subofficeId);
-		request.setAttribute("businessHallId", businessHallId);
 		PageResout pageResout=new PageResout();
 		pageResout.setCurrentPage(currentPage);
 		uimService.findForPageForShow(userName,cardNum,subofficeId,businessHallId,pageResout);
 		request.setAttribute("pageResout", pageResout);
-		request.setAttribute("userName", userName);
-		request.setAttribute("cardNum", cardNum);
-	//	return "workBill/uim/list3";
 		return "workBill/uim/list1";
 	}
+	private String 	list2(String userName,String cardNum,Integer businessHallId, Integer currentPage,HttpServletRequest request,User user)
+	{
+		List<BusinessHall> businessHalls= businessHallService.getBySubofficeId(1);
+		request.setAttribute("businessHalls", businessHalls);
+		PageResout pageResout=new PageResout();
+		pageResout.setCurrentPage(currentPage);
+		uimService.findForPageForShow(userName,cardNum,user.getSubofficeId(),businessHallId,pageResout);
+		request.setAttribute("pageResout", pageResout);
+		return "workBill/uim/list2";
+	}
 	@RequestMapping("/details")
-	public String details(int id,HttpServletRequest request)
+	public String details(int id,HttpServletRequest request,HttpSession session)
 	{
 		Map<String, Object> uim=  uimService.getDetailsByIdForShow(id);
 		request.setAttribute("uim", uim);
-		request.setAttribute("canChangeDealState", true);
-		request.setAttribute("canDelete", true);
-		request.setAttribute("canRemark", true);
+		User user = (User) session.getAttribute("user");
+		switch (user.getLimitsId()) {
+		case 1:
+			request.setAttribute("canChangeDealState", true);
+			request.setAttribute("canDelete", true);
+			request.setAttribute("canRemark", true);
+			break;
+		case 2:
+			request.setAttribute("canChangeDealState", true);
+			request.setAttribute("canDelete", true);
+			break;
+
+		default:			
+		}
 		return "workBill/uim/details";
 	}
 	@RequestMapping("/delete")	
-	public String delete(int id,HttpServletRequest request)
+	public String delete(int id,HttpServletRequest request,HttpSession session)
 	{
+		User user = (User) session.getAttribute("user");
+		if(user.getLimitsId()!=1)
+		{
+			return "";
+		}
 		uimService.remove(id);
 		request.setAttribute("backUrl", "uim/list.do");
 		return "util/optSuccess";
 	}
 	@RequestMapping("/changeDealState")	
-	public String changeDealState(int id,int dealState,HttpServletRequest request)
+	public String changeDealState(int id,int dealState,HttpServletRequest request,HttpSession session)
 	{
+		Uim uim =uimService.getById(id);
+		if(!isHaveLimits(uim,session))
+		{
+			return"";
+		}
 		uimService.changeDealState(id,dealState);
 		request.setAttribute("backUrl", "uim/list.do");
 		return "util/optSuccess";
@@ -132,9 +192,13 @@ public class UimController {
 		return "util/optSuccess";
 	}	
 	@RequestMapping("/downFile")
-	public void downFile(int id,String type,HttpServletResponse response)
+	public void downFile(int id,String type,HttpServletResponse response,HttpSession session)
 	{
 		Uim uim= uimService.getById(id);
+		if(!isHaveLimits(uim,session))
+		{
+			return;
+		}
 		if(uim!=null)
 		{
 			String fileName;
@@ -159,5 +223,24 @@ public class UimController {
 			downName=fileName.substring(fileName.indexOf("_")+1);
 			SysUtil.downFile(filePath+fileName,downName, response);
 		}
+	}
+	private boolean isHaveLimits(Uim uim,HttpSession session)
+	{
+		User user = (User) session.getAttribute("user");
+		if(user.getLimitsId()==2)
+		{
+			if(uim.getSubofficeId()!=user.getSubofficeId())
+			{
+				return false;
+			}
+		}
+		else if(user.getLimitsId()==3)
+		{
+			if(user.getBusinessHallId()!=user.getBusinessHallId())
+			{
+				return false;
+			}
+		}
+		return true;
 	}
 }
